refactor: Modularize getDayLabel function
This commit refactors the getDayLabel function by modularizing its logic into a separate component.
